Create EnderDragonMock
