+++ src/main/java/org/mockbukkit/mockbukkit/entity/EnderDragonMock.java
@@ -35,105 +35,7 @@ public class EnderDragonMock extends AbstractBossMock implements EnderDragon
 	public EnderDragonMock(@NotNull ServerMock server, @NotNull UUID uuid)
 	{
 		super(server, uuid, "Ender Dragon");
-		battle = new DragonBattle()
-		{
-			@Override
-			public @Nullable EnderDragon getEnderDragon()
-			{
-				return EnderDragonMock.this;
-			}
-
-			@Override
-			public @NotNull BossBar getBossBar()
-			{
-				return EnderDragonMock.this.bossBarMock;
-			}
-
-			@Override
-			public @Nullable Location getEndPortalLocation()
-			{
-				return null;
-			}
-
-			@Override
-			public boolean generateEndPortal(boolean generatesEndPortal)
-			{
-				return false;
-			}
-
-			@Override
-			public boolean hasBeenPreviouslyKilled()
-			{
-				return false;
-			}
-
-			@Override
-			public void setPreviouslyKilled(boolean previouslyKilled)
-			{
-				throw new UnimplementedOperationException();
-			}
-
-			@Override
-			public void initiateRespawn()
-			{
-				throw new UnimplementedOperationException();
-			}
-
-			@Override
-			public boolean initiateRespawn(@Nullable Collection<EnderCrystal> enderCrystalCollection)
-			{
-				return false;
-			}
-
-			@NotNull
-			@Override
-			public RespawnPhase getRespawnPhase()
-			{
-				throw new UnimplementedOperationException();
-			}
-
-			@Override
-			public boolean setRespawnPhase(@NotNull DragonBattle.RespawnPhase respawnPhase)
-			{
-				return false;
-			}
-
-			@Override
-			public void resetCrystals()
-			{
-				throw new UnimplementedOperationException();
-			}
-
-			@Override
-			public int getGatewayCount()
-			{
-				return 0;
-			}
-
-			@Override
-			public boolean spawnNewGateway()
-			{
-				return false;
-			}
-
-			@Override
-			public void spawnNewGateway(@NotNull Position position)
-			{
-				throw new UnimplementedOperationException();
-			}
-
-			@Override
-			public @NotNull @Unmodifiable List<EnderCrystal> getRespawnCrystals()
-			{
-				return List.of();
-			}
-
-			@Override
-			public @NotNull @Unmodifiable List<EnderCrystal> getHealingCrystals()
-			{
-				return List.of();
-			}
-		};
+		battle = new DragonBattleMock(this);
 		this.setMaxHealth(200.0D);
 		this.setHealth(200.0D);
 	}
